requirements.txt
Flask==2.2.2
Flask_RESTful==0.3.9
Flask_Script==2.0.5
Flask_SQLAlchemy==3.0.2
Flask-Migrate==4.0.4
PyJWT==2.6.0
pytest==7.2.1
requests==2.27.1
SQLAlchemy_Utils==0.39.0
python-dotenv==0.21.1
nameof==0.0.1
Flask-Mail==0.9.1
email_validator==1.3.1
celery==5.2.7